Explicitly set up the head branch after clone
Before this patch we rely on the behavior that `git checkout X` would
set up a remote tracking branch matching branch of the same name on the
only remote. But sometimes such a branch already exists, and chaos
ensues.

I hit a rare edge case today when the "master" branch already exists in
the bare repo and the cloned repo has its HEAD set to a local branch
"master" tracking fast-clone/master ...